refactored setup_scanner to put all the code for generating signals for the actual trading strategy together in a function, which is now in the strategies module. this will make it much easier to re-use the rest of the code when i write new strategies and makes it more readable. i saved the code as it was before refactoring in the file 'setup_scanner_old.py'
there was a bug with the clear_stop function where if there was not stop order set,
it would cause an exception and not sell the asset. fixed it by adding a condition
to only try to clear stops if there are any to clear.

setup_scanner had a bug when update_ohlc was called on very short dataframes (2 rows or less),
so i put update_ohlc inside an if block which checks that the df is longer than 2 before it
runs update_ohlc.